requirements.txt
ultralytics
supervision
opencv-python
tqdm
ffmpegcv
psycopg2-binary
pytz
boto3
python-dotenv
